Adding 2x2 exp/log convenience methods
