+++ include/common/qrack_types.hpp
@@ -12,6 +12,8 @@
 
 #pragma once
 
+#include <memory>
+
 #define bitLenInt uint8_t
 #define bitCapInt uint64_t
 #define bitsInByte 8
@@ -39,4 +41,7 @@
 
 namespace Qrack {
 typedef std::shared_ptr<complex> BitOp;
-}
+
+void exp(complex* matrix2x2, complex* outMatrix2x2);
+void log(complex* matrix2x2, complex* outMatrix2x2);
+} // namespace Qrack
